# [3.x] Faster queue free