@lawnjelly 2022-11-24
Made me wonder what the maximum number of children might be! Not sure this is correct but :grinning:

The group can be a maximum of 536,870,911 .. (29 bits)
We have 31 bits to play with before overflow to negative value (maximum value 2147483647 which is `INT32_MAX`) ..

so I think we'd need 2147483647 - 536870911 to overflow:
1,610,612,736 children?

So I think that's enough for 1 and a half billion children of a node. (It could happen!)