@RandomShaper 2022-11-24
I'm wondering if we should consider this a compat breaking change since it can affect observable behavior (e.g., user scripts handling `NOTIFICATION_PREDELETE` that for some reason realy on the current order of deletion). In that case it may be good that in 3.x there's a project setting to opt-in sorting (I think if the "queue" is not sorted, the behavior is as formerly, right?). Such a setting could be disabled by default but enabled for new projects, as we've done with other compat breaking but expected to be universally wanted changes.