@akien-mga 2022-11-25
> > I'm wondering if we should consider this a compat breaking change since it can affect observable behavior (e.g., user scripts handling `NOTIFICATION_PREDELETE` that for some reason realy on the current order of deletion). In that case it may be good that in 3.x there's a project setting to opt-in sorting (I think if the "queue" is not sorted, the behavior is as formerly, right?). Such a setting could be disabled by default but enabled for new projects, as we've done with other compat breaking but expected to be universally wanted changes.
> 
> Sure I can put this in. I would strongly suggest that any sort of guarantee of order should be discouraged - I'll add this in the classref for `queue_free()`. Without a written guarantee I wouldn't expect wise users to rely on this, but you are correct it is at least possible. Although if there are no reports we could consider removing the option after a couple of betas.

Since we're speculating here, I prefer going with the "ask for forgiveness" approach. Change the behavior, see if anyone complains, and consider adding an option for those users if they exist.