@lawnjelly 2022-11-25
> Since we're speculating here, I prefer going with the "ask for forgiveness" approach. Change the behavior without opt-out, see if anyone complains, and consider adding an option for those users if they exist. Otherwise we're just adding configuration options for things which may not be needed.
> 
> Betas don't have to be production ready, so if this behavior change is a showstopper for some users, they should report it after testing the beta and stick to using 3.5.x until their feedback is addressed in later 3.6 betas.

I actually agree here... and I'd just added the project setting lol. I'll change it back, but keep a comment in the `queue_free` classref to mention that the order is not guaranteed. :+1: