@RandomShaper 2022-11-25
> I'm not super familiar with hash set - what advantage are you thinking in terms of hash set over sorting? I'm assuming the standard Godot sort is something like qsort (i.e. in place and not using dynamic allocation) .. I haven't looked in detail.

To be honest, it's hard to tell for sure without benchmarking, but my idea with a sorted container (it'd be `RBSet` actually) there may be some code complexity or maybe performance advantages. Just an idea if you wish to try and compare.