@eimfach 2023-01-09
+1 for this. I am developing a hex turn-based strategy game with lots of tiles and need a fast `queue_free` and/or `free` for releasing tiles out of view. Would be very happy to see this fix soon ! :)