@akien-mga 2023-01-10
We discussed this yesterday with @reduz and @lawnjelly.
@reduz wants to explore an alternative solution for 4.x, so we'll hold off merging the `master` PR for now, but for `3.x` we agreed to merge this approach.